+++ android-agent/src/main/kotlin/io/opentelemetry/android/agent/dsl/SessionConfiguration.kt
@@ -16,7 +16,7 @@ import kotlin.time.Duration.Companion.minutes
 @OpenTelemetryDslMarker
 class SessionConfiguration internal constructor() {
     /**
-     * The maximum duration which a session can remain open in the background before it
+     * The maximum duration which a session can remain open without meaningful activity before it
      * automatically expires.
      */
     var backgroundInactivityTimeout: Duration = 15.minutes